Merge retrofit-teste into master
Retrofit implementation

The app is now running with most of the 1st sprint specifications.
Features developed during this first sprint:

* Integration between the app and back-end service via Retrofit
https://trello.com/c/Q4xnkgsK/46-integrar-o-aplicativo-com-o-servi%C3%A7o-de-back-end

* Recover data from service and handle it.
https://trello.com/c/OO3BVHg0/60-recuperar-os-destaques-do-servi%C3%A7o-back-end
https://trello.com/c/8gt9VFib/64-tratar-os-dados-recuperados-do-servi%C3%A7o-back-end

* Display highlights in the homescreen
https://trello.com/c/4D06e3GL/65-exibir-os-destaques-na-home-screen

* Clickable highlights cards.
https://trello.com/c/vONXapOM/66-tornar-os-destaques-clic%C3%A1veis

* Redirect the user to the Instagram app after clicking on any highlight card
https://trello.com/c/X3zhOI1y/67-redirecionar-os-destaques-%C3%A0s-postagens-do-instagram

* Handle connection errors
https://trello.com/c/glGgPPQ0/62-tratar-erros-de-conex%C3%A3o

* Some software improvements:
 - SwipeRefreshLayout is now working.
 - Most text resources are now at strings.xml.
 - Changes in the models due to database changes.

The user is now capable of accessing the highlights on the homescreen.
The highlights were previously registered on our database via an web app and
loaded into Cards to be displayed in a 2x2 grid layout.
Upon clicking any of the cards, the user will be redirected to the highlight's instagram post
page on the Instagram app.
If the user has connection problems while fetching data from the servers, a toast message
will be displayed warning about connection issues.